refactor: move server/git tests into test/server/git
