+++ test/server/git/issues.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { normalizeIssue, ISSUE_LIMIT } from "./issues";
+import { normalizeIssue, ISSUE_LIMIT } from "../../../server/git/issues";
 
 describe("normalizeIssue", () => {
   it("normalizes a gh issue json object", () => {
