+++ test/server/git/worktree-pr.spec.ts
@@ -3,8 +3,8 @@ import { mkdtempSync, rmSync, writeFileSync, realpathSync } from "node:fs";
 import { execFileSync } from "node:child_process";
 import { tmpdir } from "node:os";
 import path from "node:path";
-import { createWorktree } from "./worktrees.js";
-import { compareUrl, pushWorktree, createOrOpenPR } from "./worktree-pr.js";
+import { createWorktree } from "../../../server/git/worktrees.js";
+import { compareUrl, pushWorktree, createOrOpenPR } from "../../../server/git/worktree-pr.js";
 
 describe("compareUrl", () => {
   it("builds the GitHub compare/open-PR url, keeping the branch slash raw", () => {
